Remove ndp.go and add icmpv6.go
